Details

+++ Timer.cpp
@@ -28,7 +28,7 @@ void Timer::Tick()
 	auto delta = t1 - t0;
 	t0 = t1;
 
-	elapsedTime = delta.count();
+	elapsedTime = static_cast<double>(delta.count());
 	totalTime += elapsedTime;
 }
